pyproject.toml
[project]
name = "wisecofre"
version = "1.0.0"
description = "Gerenciador de senhas e segredos para times — self-hosted, E2E encrypted"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "django>=5.2,<7.0",
    "djangorestframework>=3.15",
    "django-environ>=0.12",
    "django-filter>=25.1",
    "djangorestframework-simplejwt>=5.4",
    "drf-spectacular>=0.28",
    "django-allauth>=65.0",
    "django-otp>=1.5",
    "pyotp>=2.9",
    # "django-auth-ldap>=5.1",  # Requires C build tools; install in Docker
    "python-gnupg>=0.5",
    "cryptography>=44.0",
    "celery[redis]>=5.4",
    "django-celery-beat>=2.7",
    "django-anymail>=12.0",
    "structlog>=25.1",
    "psycopg[binary]>=3.2",
    "redis>=5.2",
    "django-cors-headers>=4.6",
    "django-ratelimit>=4.1",
    "gunicorn>=23.0",
    "whitenoise>=6.8",
    "django-storages[s3]>=1.14",
    "boto3>=1.34",
    "Pillow>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "factory-boy>=3.3",
    "faker>=33.0",
    "ruff>=0.11",
    "django-debug-toolbar>=5.0",
    "ipython>=8.30",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py"]
addopts = "-v --tb=short --strict-markers"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pytest-playwright>=0.7.2",
]
